+++ test/test-db-Project.js
@@ -1,27 +1,33 @@
-/*jshint expr: true*/
 'use strict';
 
-var db = require('../lib/db');
-var chai = require('chai');
+const db = require('../lib/db');
+const chai = require('chai');
 
 chai.should();
 
-describe('db/Project', function() {
+const projectType = 'dt';
+const projectId = 'di';
+const configuration = {foo: 'bar'};
 
+describe('db/Project', function() {
     beforeEach(function(done) {
-        db.projects._model.collection.drop().then(function() {
-            done();
-          }).catch(function() {
-            // ignore errors
-          });
+        const checkForDb = function() {
+          if (!db.isConnected()) {
+            return setTimeout(checkForDb, 100);
+          }
+
+          db.projects._model.collection.drop().then(function() {
+              done();
+            }).catch(function() {
+              done();
+            });
+          };
+
+        checkForDb();
       });
 
     it('should save project', function(done) {
-        var projectType = 'pt';
-        var projectId = 'pi';
-        var configuration = {foo: 'bar'};
-
-        var step1 = function() {
+        const step1 = function() {
             db.projects.save(projectType, projectId, configuration,
               function(isSaved) {
                 isSaved.should.not.be.false;
@@ -29,17 +35,14 @@ describe('db/Project', function() {
               });
           };
 
-        var step2 = function() {
+        const step2 = function() {
             db.projects._model.find({
                 project_type: projectType,
-                project_id: projectId
+                project_id: projectId,
               }, function(err, projects) {
                 projects.should.be.a('array');
                 projects.length.should.equal(1);
-
-                var project = projects[0];
-                project.configuration.should.be.a('object');
-                project.configuration.foo.should.equal(configuration.foo);
+                projects[0].configuration.should.deep.equal(configuration);
 
                 done();
               });
@@ -49,13 +52,13 @@ describe('db/Project', function() {
       });
 
     it('should save apn', function(done) {
-        var bundleId = 'bi';
-        var tokenKey = 'tk';
-        var tokenKeyId = 'tki';
-        var tokenTeamId = 'tti';
-        var production = true;
+        const bundleId = 'bi';
+        const tokenKey = 'tk';
+        const tokenKeyId = 'tki';
+        const tokenTeamId = 'tti';
+        const production = true;
 
-        var step1 = function() {
+        const step1 = function() {
             db.projects.saveApn(
               bundleId,
               tokenKey,
@@ -69,15 +72,15 @@ describe('db/Project', function() {
             );
           };
 
-        var step2 = function() {
+        const step2 = function() {
             db.projects._model.find({
                 project_type: 'apn',
-                project_id: bundleId
+                project_id: bundleId,
               }, function(err, projects) {
                 projects.should.be.a('array');
                 projects.length.should.equal(1);
 
-                var project = projects[0];
+                const project = projects[0];
                 project.configuration.should.be.a('object');
                 project.configuration.token.should.be.a('object');
                 project.configuration.token.key.should.equal(tokenKey);
@@ -93,25 +96,25 @@ describe('db/Project', function() {
       });
 
     it('should save gcm', function(done) {
-        var packageId = 'pi';
-        var apiKey = 'ak';
+        const packageId = 'pi';
+        const apiKey = 'ak';
 
-        var step1 = function() {
+        const step1 = function() {
             db.projects.saveGcm(packageId, apiKey, function(isSaved) {
                 isSaved.should.not.be.false;
                 step2();
               });
           };
 
-        var step2 = function() {
+        const step2 = function() {
             db.projects._model.find({
                 project_type: 'gcm',
-                project_id: packageId
+                project_id: packageId,
               }, function(err, projects) {
                 projects.should.be.a('array');
                 projects.length.should.equal(1);
 
-                var project = projects[0];
+                const project = projects[0];
                 project.configuration.should.be.a('object');
                 project.configuration.api_key.should.equal(apiKey);
 
@@ -123,11 +126,11 @@ describe('db/Project', function() {
       });
 
     it('should save wns', function(done) {
-        var packageId = 'pi';
-        var clientId = 'ci';
-        var clientSecret = 'cs';
+        const packageId = 'pi';
+        const clientId = 'ci';
+        const clientSecret = 'cs';
 
-        var step1 = function() {
+        const step1 = function() {
             db.projects.saveWns(packageId, clientId, clientSecret,
               function(isSaved) {
                 isSaved.should.not.be.false;
@@ -135,15 +138,15 @@ describe('db/Project', function() {
               });
           };
 
-        var step2 = function() {
+        const step2 = function() {
             db.projects._model.find({
                 project_type: 'wns',
-                project_id: packageId
+                project_id: packageId,
               }, function(err, projects) {
                 projects.should.be.a('array');
                 projects.length.should.equal(1);
 
-                var project = projects[0];
+                const project = projects[0];
                 project.configuration.should.be.a('object');
                 project.configuration.client_id.should.equal(clientId);
                 project.configuration.client_secret.should.equal(clientSecret);
@@ -156,24 +159,22 @@ describe('db/Project', function() {
       });
 
     it('should update project', function(done) {
-        var projectType = 'dt';
-        var projectId = 'di';
-        var configuration = {foo: 'bar'};
-        var configuration2 = {bar: 'foo'};
-        var theProject = null;
+        const configuration2 = {bar: 'foo'};
+        let theProject = null;
 
-        var init = function() {
+        const init = function() {
             db.projects._model.create({
                 project_type: projectType,
                 project_id: projectId,
-                configuration: configuration
+                configuration: configuration,
               }, function(err, project) {
+                project.should.not.be.null;
                 theProject = project;
                 step1();
               });
           };
 
-        var step1 = function() {
+        const step1 = function() {
             db.projects.save(projectType, projectId, configuration2,
               function(isSaved) {
                 isSaved.should.not.be.false;
@@ -181,7 +182,7 @@ describe('db/Project', function() {
               });
           };
 
-        var step2 = function() {
+        const step2 = function() {
             db.projects._model.findById(theProject._id,
               function(err, project) {
                 project.configuration.should.has.all.keys('foo', 'bar');
@@ -200,22 +201,19 @@ describe('db/Project', function() {
       });
 
     it('should return project', function(done) {
-        var projectType = 'dt';
-        var projectId = 'di';
-        var configuration = {foo: 'bar'};
-        var now = Date.now();
+        const now = Date.now();
 
-        var init = function() {
+        const init = function() {
             db.projects._model.create({
                 project_type: projectType,
                 project_id: projectId,
-                configuration: configuration
+                configuration: configuration,
               }, function() {
                 step1();
               });
           };
 
-        var step1 = function() {
+        const step1 = function() {
             db.projects.findProject(projectType, projectId, function(project) {
                 project.should.be.a('object');
                 project.project_type.should.equal(projectType);
@@ -232,25 +230,21 @@ describe('db/Project', function() {
       });
 
     it('should return project configuration', function(done) {
-        var projectType = 'dt-config';
-        var projectId = 'di-config';
-        var configuration = {foo: 'bar'};
-
-        var init = function() {
+        const init = function() {
             db.projects._model.create({
                 project_type: projectType,
                 project_id: projectId,
-                configuration: configuration
+                configuration: configuration,
               }, function() {
                 step1();
               });
           };
 
-        var step1 = function() {
+        const step1 = function() {
             db.projects.findConfig(projectType, projectId,
               function(projectConfig) {
                 projectConfig.should.be.a('object');
-                projectConfig.foo.should.equal(configuration.foo);
+                projectConfig.should.deep.equal(configuration);
 
                 done();
               });
